Added a bank table to the database to track the balance in your bank

+++ src/main/java/com/thevault/beans/Entertainment.java
@@ -31,7 +31,7 @@ public class Entertainment {
 
 	@Column(name="AMOUNT")
 	@Qualifier("amount")
-	private int amount;
+	private double amount;
 	
 	@Column(name="DESCRIPTION")
 	@Qualifier("description")
@@ -64,11 +64,11 @@ public class Entertainment {
 		this.entertainmentId = entertainmentId;
 	}
 
-	public int getAmount() {
+	public double getAmount() {
 		return amount;
 	}
 
-	public void setAmount(int amount) {
+	public void setAmount(double amount) {
 		this.amount = amount;
 	}
